+++ src/Generators/Middleware.php
@@ -22,6 +22,14 @@ final class Middleware
 
         PHP;
 
+        if (!\file_exists('./src/modules/Middleware')) {
+            echo
+            '---------------------------------------------------' . \PHP_EOL .
+            'WARNING: dir [./src/modules/Middleware] not exists!' . \PHP_EOL .
+            '---------------------------------------------------' .\PHP_EOL;
+            exit(1);
+        }
+
         \file_put_contents("./src/modules/Middleware/{$name}.php", $content);
         echo
         '--------------------------------------------------------------------' . \PHP_EOL .
