+++ src/Generators/Model.php
@@ -24,6 +24,14 @@ final class Model
 
         PHP;
 
+        if (!\file_exists('./src/modules/Models')) {
+            echo
+            '-----------------------------------------------' . \PHP_EOL .
+            'WARNING: dir [./src/modules/Models] not exists!' . \PHP_EOL .
+            '-----------------------------------------------' .\PHP_EOL;
+            exit(1);
+        }
+
         \file_put_contents("./src/modules/Models/{$name}.php", $content);
         echo
         '------------------------------------------------------------' . \PHP_EOL .
